AddNodePanel component mockup added on ACL for the Restricted Mode
